+++ msvsphere/rpmqc/cli.py
@@ -1,5 +1,7 @@
+from enum import IntEnum
 import argparse
 import sys
+import traceback
 
 from . import __version__
 from .config import Config
@@ -7,17 +9,43 @@ from .file_utils import normalize_path
 from .runner import run_rpm_inspections
 
 
-def init_arg_parser() -> argparse.ArgumentParser:
+class ExitCodes(IntEnum):
+
+    """
+    The program exit codes (pytest-compatible).
+
+    References:
+        https://pytest.org/en/latest/reference/reference.html#pytest.ExitCode
+    """
+
+    PASSED = 0               # all tests were successful
+    FAILED = 1               # some tests failed
+    INTERRUPTED = 2          # the program was terminated by a user
+    INTERNAL_ERROR = 3       # unexpected error happened
+    USAGE_ERROR = 4          # command line arguments/config error
+    NO_TESTS_FOUND = 5       # no tests were found
+
+
+class ArgParser(argparse.ArgumentParser):
+
+    """
+    ArgumentParser implementation that uses a custom exit code on usage error.
+    """
+
+    def error(self, message: str):
+        self.print_usage(sys.stderr)
+        self.exit(ExitCodes.USAGE_ERROR, f'{self.prog}: {message}\n')
+
+
+def init_arg_parser() -> ArgParser:
     """
     Initializes a command line argument parser.
 
     Returns:
         Command line arguments parser.
     """
-    parser = argparse.ArgumentParser(
-        prog='rpmqc',
-        description='RPM packages quality control tool'
-    )
+    parser = ArgParser(prog='rpmqc',
+                       description='RPM packages quality control tool')
     parser.add_argument('-c', '--config', help='configuration file path',
                         required=True)
     parser.add_argument('--version', action='version',
@@ -37,5 +65,14 @@ def main():
     arg_parser = init_arg_parser()
     args = arg_parser.parse_args(sys.argv[1:])
     cfg = Config(args.config)
-    if args.command == 'inspect-rpm':
-        run_rpm_inspections(cfg, args.rpm_path)
+    success = False
+    try:
+        if args.command == 'inspect-rpm':
+            success = run_rpm_inspections(cfg, args.rpm_path)
+    except KeyboardInterrupt:
+        sys.stderr.write('rpmqc: interrupted by user\n')
+        sys.exit(ExitCodes.INTERRUPTED)
+    except Exception:
+        traceback.print_exc()
+        sys.exit(ExitCodes.INTERNAL_ERROR)
+    sys.exit(ExitCodes.PASSED if success else ExitCodes.FAILED)