+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Exit with a non-zero exit code if some tests failed.
+- Use pytest-compatible exit codes.
+
+
 ## [0.0.2] - 2023-10-28
 
 ### Added